#96 pickup heart regenerate health

+++ Silesian Undergrounds/Silesian Undergrounds/Engine/Item/Heart.cs
@@ -7,7 +7,7 @@ using Silesian_Undergrounds.Engine.Collisions;
 namespace Silesian_Undergrounds.Engine.Item {
     public class Heart : PickableItem{
 
-        private int healtRegenerationValue = 0;
+        private int liveRegenerationValue = 25;
 
         public Heart(Texture2D texture, Vector2 position, Vector2 size, int layer, Scene.Scene scene) : base(texture, position, size, layer, scene)
         {
@@ -19,7 +19,13 @@ namespace Silesian_Undergrounds.Engine.Item {
         {
             if (obj is Player)
             {
-                this.scene.DeleteObject(this);
+                Player pl = (Player)obj;
+                if (pl.MaxLiveValue > pl.LiveValue)
+                {
+                    pl.RefilLive(this.liveRegenerationValue);
+                    this.scene.DeleteObject(this);
+                }
+
             }
         }
     }
